backend/requirements.txt
Django==6.0.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
django-filter==25.2
pillow==12.1.0
boto3==1.42.30
django-storages==1.14.6
gunicorn